fix: move Assign preview caption out of the screens file-size ceiling.
The Development scope list was 303 physical lines, which failed repo.app_features_screens_lib_physical_file_size.
